leetcode/src/test/java/com/kndykndy/leetcode/p202_HappyNumberTest.java
package com.kndykndy.leetcode;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

import com.kndykndy.leetcode.p202_HappyNumber.Solution;
import com.kndykndy.leetcode.p202_HappyNumber.Solution1;
import java.util.Collections;
import java.util.List;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

class p202_HappyNumberTest {

    private List<Solution> solutions;

    @BeforeEach
    void setUp() {
        solutions = Collections.singletonList(new Solution1());
    }

    @Test
    void test1() {
        for (Solution solution : solutions) {
            assertTrue(solution.isHappy(19));
            assertFalse(solution.isHappy(25));
        }
    }
}
